+++ test/month_picker_dialog_test.dart
@@ -49,7 +49,7 @@ void main() {
           yearTextStyle: TextStyle(
             fontWeight: FontWeight.bold,
           ),
-          selectedDatePadding: 1,
+          selectedDateRadius: 1,
         ),
       ),
     );
@@ -66,7 +66,7 @@ void main() {
     expect(controller.localLastDate, DateTime(2027));
     controller.firstPossibleMonth(2120);
     expect(controller.selectedDate, DateTime(2120, 2));
-    expect(controller.monthPickerDialogSettings.buttonsSettings.selectedDatePadding, 1);
+    expect(controller.monthPickerDialogSettings.buttonsSettings.selectedDateRadius, 1);
     expect(controller.monthPickerDialogSettings.dialogSettings.forcePortrait,
         true);
     expect(controller.rangeMode, true);